[#217] The `.active` file is now versioned and contains a slightly different commit:
It used to contain SHA1 of the first VersionPress commit (its initialization commit) which could be done only after that initial commit has been created. Now that the `.active` file is versioned as well, this would mean a need to make another commit. So instead, the `.active` file now contains one commit before the initialization commit (or an empty string if the initialization commit is the first one ever in the repo) and initialization can still be a single commit.