Tricky issue: there's probably a Symfony Process bug that sometimes converts an empty string argument to a single quote character. It's quite strange, see: https://github.com/symfony/symfony/issues/23455.
The fix is to try to ensure that the important $args are the last ones. For example, if the main $args are `"DB_COLLATE"` and `""`, all the other args like `--color` and `--allow-root` are now prepended so that the final command reads e.g. `wp vp update-config --color --allow-root "DB_COLLATE" ""`. In this order, everything works.