Bye bye, Gulp. Everything is now scripted via npm scripts directly, or in more complex cases, via TypeScript files in the `./scripts` directory.
Why the change?

Gulp used to be popular (and progressive!) when we first started with VersionPress but today, it feels like an unnecessary abstraction. In our specific case, it was probably causing strange issues with nested `npm run`s – hard to tell as some dependencies were a couple of years old / unmaintained. Also, in the past, we've had trouble generating ZIP files reliably and Gulp itself doesn't seem to be that actively developed anyway.

So I rewrote the functionality to a simpler and more straightforward code. Less dependencies = win.