Break ThemesTest again but in a better way :)
The fix in fcde513f91a6aaa1c9b654b286eeb898f905ceff has been reverted as that will be better dealt with in a separate PR (I'm not entirely sure the fix was correct and don't want to "hide" it in this already huge PR).

On the plus side, ThemesTest now continue running even after `switchingThemeCreatesThemeSwitchAction()` fails.