Run cli as user 33:33 to match www-data from versionpress/wordpress:php7.2-apache
versionpress/wordpress:php7.2-apache is based on wordpress:php7.2-apache which is based on debian.

On Debian, www-data:www-data is 33:33
On Alpine (which is what cli is based on), www-data:www-data is 82:82
Therefore, when Alpine and Debian interact, one has to change to match the other.

See https://github.com/docker-library/wordpress/issues/256 for details