Run CLI image as UID 33 (Debian's "www-data")
Alpine's www-data is UID 82. To make the two images work together, e.g., writes to `/var/www/html` by both Apache and WP-CLI, this image is customized to use UID 33 explicitly (it's user 'xfs' on Alpine Linux).

As a result, it's no longer necessary to set the user in docker-compose YAML.